# backend/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"DevBank API","description":"REST API for DevBank backend services, providing endpoints for customer management, accounts, transactions, and more.","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Card","description":"Endpoints for managing cards"},{"name":"Payment","description":"Endpoints for managing payments"},{"name":"Transaction","description":"Endpoints for managing transactions and loan payments"},{"name":"Report","description":"Endpoints for generating account and card transaction reports"},{"name":"Account","description":"Endpoints for managing accounts"},{"name":"Customer","description":"Endpoints for managing customers"},{"name":"Loan","description":"Endpoints for managing loans"}],"paths":{"/bia/customer/{customerId}":{"get":{"tags":["Customer"],"summary":"customerByID","description":"Retrieves customer details by customer ID","operationId":"customerById","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Customer found","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Customer not found"}}},"put":{"tags":["Customer"],"summary":"customerUpdate","description":"Updates customer details by customer ID","operationId":"customerUpdate","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerUpdate"}}},"required":true},"responses":{"200":{"description":"Customer updated successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Customer not found"}}},"delete":{"tags":["Customer"],"summary":"customerDelete","description":"Deletes a customer by ID","operationId":"customerDelete","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Customer deleted successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Customer not found"}}}},"/bia/cards/{id}":{"put":{"tags":["Card"],"summary":"cardUpdate","description":"Updates card details by card ID","operationId":"cardUpdate","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreditUpdate"}}},"required":true},"responses":{"200":{"description":"Card updated successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Card not found"}}},"delete":{"tags":["Card"],"summary":"cardDelete","description":"Deletes a card by its ID","operationId":"cardDelete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Card deleted successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Card not found"}}}},"/bia/account/{accountNumber}":{"get":{"tags":["Account"],"summary":"getAccountByNumber","description":"Retrieves account details by account number","operationId":"getAccountByNumber","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Account found","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Account not found"}}},"put":{"tags":["Account"],"summary":"accountUpdate","description":"Updates the details of an existing account","operationId":"accountUpdate","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountUpdate"}}},"required":true},"responses":{"200":{"description":"Account updated successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Account not found"}}},"delete":{"tags":["Account"],"summary":"accountDelete","description":"Deletes an account by its number","operationId":"accountDelete","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Account deleted successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Account not found"}}}},"/bia/transactions/{originAccountNumber}":{"post":{"tags":["Transaction"],"summary":"createTransaction","description":"Performs a transaction from the origin account to the target account","operationId":"createTransaction","parameters":[{"name":"originAccountNumber","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionRequest"}}},"required":true},"responses":{"200":{"description":"Transaction completed successfully","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid transaction data"}}}},"/bia/transactions/loanPayments/{loanPaymentsId}":{"post":{"tags":["Transaction"],"summary":"transactionAddLoanPayments","description":"Adds a transaction as a loan payment","operationId":"transactionAddLoanPayments","parameters":[{"name":"loanPaymentsId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionRequest"}}},"required":true},"responses":{"200":{"description":"Loan payment added successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Loan payment not found"}}}},"/bia/loans/{customerId}":{"post":{"tags":["Loan"],"summary":"createLoan","description":"Creates a new loan for a specific customer","operationId":"createLoan","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoanRequest"}}},"required":true},"responses":{"201":{"description":"Loan created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}},"400":{"description":"Invalid input data"}}}},"/bia/customer":{"post":{"tags":["Customer"],"summary":"createCustomer","description":"Registers a new costumer in the BIA system","operationId":"createCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerRequest"}}},"required":true},"responses":{"200":{"description":"Customer successfully created","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid input data"}}}},"/bia/cards/debit":{"post":{"tags":["Card"],"summary":"Add a debit card purchase","description":"Registers a new purchase using the debit card, deducting the amount from the linked account.","operationId":"addDebitBuying","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardPaymentsRequest"}}},"required":true},"responses":{"200":{"description":"Debit card purchase added successfully","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid purchase details","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Card or account not found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bia/cards/credit":{"post":{"tags":["Card"],"summary":"Add a credit card purchase","description":"Registers a new purchase on the credit card and updates the card balance.","operationId":"addCreditBuying","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardPaymentsRequest"}}},"required":true},"responses":{"200":{"description":"Credit card purchase added successfully","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid purchase details","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Card not found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bia/cards/add/{accountNumber}":{"post":{"tags":["Card"],"summary":"Create a new card for an account","description":"Creates a new card associated with the specified account number.","operationId":"createCard","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreditRequest"}}},"required":true},"responses":{"201":{"description":"Card created successfully","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid input data","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Account not found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bia/account/{customerId}":{"post":{"tags":["Account"],"summary":"createAccount","description":"Creates a new bank account for an existing customer","operationId":"createAccount","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountRequest"}}},"required":true},"responses":{"200":{"description":"Account sucessfully created","content":{"application/json":{}}},"400":{"description":"Invalid input data"}}}},"/bia/account/balanceOut/{accountNumber}":{"post":{"tags":["Account"],"summary":"accountCashOut","description":"Withdraws the specified amount from the customer's account.","operationId":"accountCashOut","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountUpdate"}}},"required":true},"responses":{"202":{"description":"Withdrawal successful","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid input data","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bia/account/balanceIn/{accountNumber}":{"post":{"tags":["Account"],"summary":"accountDeposit","description":"Deposits the specified amount into the customer's account.","operationId":"accountDeposit","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountUpdate"}}},"required":true},"responses":{"202":{"description":"Deposit successful","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid input data","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/bia/transactions/{id}":{"get":{"tags":["Transaction"],"summary":"getTransactionByID","description":"Retrieves all transactions linked to a specific account","operationId":"getTransactionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Transactions retrieved successfully","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["Transaction"],"summary":"transactionsDelete","description":"Deletes a transaction by its ID","operationId":"transactionsDelete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Transaction deleted successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Transaction not found"}}}},"/bia/transactions/list":{"get":{"tags":["Transaction"],"summary":"getAllTransactions","description":"Retrieves all transactions in the system","operationId":"getAllTransactions","responses":{"200":{"description":"List of transactions retrieved successfully","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bia/transactions/account/{accountNumber}":{"get":{"tags":["Transaction"],"summary":"getTransactionsByAccountNumber","description":"Retrieves all transactions in the system","operationId":"getTransactionsByAccountNumber","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of transactions retrieved successfully","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bia/report/transaction/{accountNumber}":{"get":{"tags":["Report"],"summary":"transactionReport","description":"Retrieves the transaction report for a specific account number","operationId":"transactionReport","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Transaction report generated successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StatementResponse"}}}}},"404":{"description":"Account not found"}}}},"/bia/report/debit/{cardNumber}":{"get":{"tags":["Report"],"summary":"cardDebitReport","description":"Retrieves all debit card transactions for a specific card number","operationId":"cardDebitReport","parameters":[{"name":"cardNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Debit card report generated successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StatementResponse"}}}}},"404":{"description":"Card not found"}}}},"/bia/report/credit/{cardNumber}":{"get":{"tags":["Report"],"summary":"cardCreditReport","description":"Retrieves all credit card transactions for a specific card number","operationId":"cardCreditReport","parameters":[{"name":"cardNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Credit card report generated successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StatementResponse"}}}}},"404":{"description":"Card not found"}}}},"/bia/report/all/{accountNumber}":{"get":{"tags":["Report"],"summary":"allReports","description":"Retrieves all reports (transactions, credit, debit) for a specific account","operationId":"allReports","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Full report generated successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StatementResponse"}}}}},"404":{"description":"Account not found"}}}},"/bia/payments/loan":{"get":{"tags":["Payment"],"summary":"getLoanPaymentByID","description":"Retrieves loan payment details for the specified loan payment ID.","operationId":"getLoanPaymentById","parameters":[{"name":"loanPaymentId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Loan payment details retrieved successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Loan payment not found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bia/payments/card/{cardPaymentId}":{"get":{"tags":["Payment"],"summary":"getCardPaymentByID","description":"Retrieves card payment details for the specified card payment ID.","operationId":"getCardPaymentById","parameters":[{"name":"cardPaymentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Card payment details retrieved successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Card payment not found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bia/loans":{"get":{"tags":["Loan"],"summary":"getAllLoans","description":"Retrieves a list of all loans in the system","operationId":"getAllLoans","responses":{"200":{"description":"List of loans retrieved successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanResponse"}}}}}}}},"/bia/loans/{id}":{"get":{"tags":["Loan"],"summary":"getLoanByID","description":"Retrieves loan details by loan ID","operationId":"getLoanById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Loan found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}},"404":{"description":"Loan not found"}}}},"/bia/cards/{cardId}":{"get":{"tags":["Card"],"summary":"getCardByID","description":"Retrieves card details by card ID","operationId":"getCardById","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Card found","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Card not found"}}}},"/bia/cards/list/{accountNumber}":{"get":{"tags":["Card"],"summary":"getListAllCardByAccount","description":"Retrieves all cards linked to a specific account","operationId":"getAllCardByAccount","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of cards retrieved successfully","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bia/account/list/{customerId}":{"get":{"tags":["Account"],"summary":"getAllAccountByCustomerID","description":"Retrieves all accounts associated with a specific customer","operationId":"getAllAccountByCostumerId","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Accounts list retrieved successfully","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"CustomerUpdate":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","maxLength":2147483647,"minLength":6},"phoneNumber":{"type":"string","pattern":"\\d{10,11}"}},"required":["email","password","phoneNumber"]},"CreditUpdate":{"type":"object","properties":{"cardLimit":{"type":"number","minimum":500.00}}},"AccountUpdate":{"type":"object","properties":{"AccountType":{"type":"string","enum":["CHECKING","SAVINGS"]},"currentBalance":{"type":"number","minimum":0.00}},"required":["currentBalance"]},"TransactionRequest":{"type":"object","properties":{"amount":{"type":"number","minimum":0.01},"destinyAccountNumber":{"type":"string"}},"required":["amount","destinyAccountNumber"]},"LoanRequest":{"type":"object","properties":{"loanAmount":{"type":"number"},"interestRate":{"type":"number","minimum":0.00},"installments":{"type":"integer","format":"int32","minimum":1},"loanType":{"type":"string","enum":["PERSONAL","BUSINESS"]}},"required":["installments","interestRate","loanAmount","loanType"]},"LoanResponse":{"type":"object","properties":{"customerName":{"type":"string"},"loanAmount":{"type":"number"},"monthlyInstallment":{"type":"number"},"installments":{"type":"integer","format":"int32"}}},"CustomerRequest":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","maxLength":2147483647,"minLength":6},"birthday":{"type":"string"},"CPF":{"type":"string","pattern":"\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}"},"phoneNumber":{"type":"string","pattern":"\\d{10,11}"}},"required":["CPF","birthday","email","name","password","phoneNumber"]},"CardPaymentsRequest":{"type":"object","properties":{"cardNumber":{"type":"string"},"productName":{"type":"string"},"totalBuying":{"type":"number"},"installmentNumber":{"type":"integer","format":"int32","minimum":1}},"required":["cardNumber","installmentNumber","productName"]},"CreditRequest":{"type":"object","properties":{"cardType":{"type":"string","enum":["DEBIT","CREDIT"]},"cardNumber":{"type":"string"},"cardLimit":{"type":"number","minimum":0.00}},"required":["cardLimit","cardNumber","cardType"]},"AccountRequest":{"type":"object","properties":{"accountType":{"type":"string","enum":["CHECKING","SAVINGS"]},"currentBalance":{"type":"number","minimum":0.00}},"required":["accountType","currentBalance"]},"AuthRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"StatementResponse":{"type":"object","properties":{"type":{"type":"string"},"amount":{"type":"number"},"timeStamp":{"type":"string","format":"date"},"description":{"type":"string"}}}}}}